requirements.txt
# Core dependencies for Feetech motor control
feetech-servo-sdk>=1.0.0
pyserial>=3.5
deepdiff>=6.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Numerical computing (required for motors2 LeKiwi kinematics)
numpy>=1.24,<2.0

# Optional: DualSense PS5 controller support (for motors/ package)
pydualsense>=0.7.0

# Optional: Computer vision and scientific computing
opencv-python>=4.5.0
scipy>=1.10
matplotlib>=3.7
